requirements/testing.txt
-r base.txt
pytest==8.3.3
pytest-django==4.9.0
pytest-env==1.1.5
coverage==7.6.8
faker==33.1.0